the key concepts of store to manage the state of book list, and how the user can add a book to and remove it from their collection within an Angular component.
Defined actions to express events.
Defined two reducer functions to manage different parts of the state.
Registered the global state container that is available throughout your application.
Defined the state, as well as selectors that retrieve specific parts of the state.
Created two distinct components, as well as a service that fetches from the Google Books API.
Injected the Store and Google Books API services to dispatch actions and select the current state.